Linux対応

+++ testData20210416/Response.cpp
@@ -210,7 +210,7 @@ float ft_stof(std::string str)
 	return (ret);
 }
 
-void getAcceptLanguages(std::map<std::string, std::vector<std::string>> &AcceptLanguageMap, std::string::iterator &itr)
+void getAcceptLanguages(std::map<std::string, std::vector<std::string> > &AcceptLanguageMap, std::string::iterator &itr)
 {
 	std::string LanguageRange;
 	std::string qValue = "1";
@@ -304,7 +304,7 @@ void setResponseMap(std::map<int, std::string> &ResponseMap)
 	ResponseMap[511] = "Network Authentication";
 }
 
-void getAcceptCharset(std::map<std::string, std::vector<std::string>> &AcceptCharsetMap, std::string::iterator &itr)
+void getAcceptCharset(std::map<std::string, std::vector<std::string> > &AcceptCharsetMap, std::string::iterator &itr)
 {
 	std::string Charset;
 	std::string qValue = "1";
@@ -425,10 +425,10 @@ bool Response::isMatchAcceptLanguageFromat(std::string src)
 	return (true);
 }
 
-std::map<std::string, std::vector<std::string>> Response::parseAcceptLanguage(std::string src)
+std::map<std::string, std::vector<std::string> > Response::parseAcceptLanguage(std::string src)
 {
 	//ここに入ってくる時点で、Accept-Languageの形式は満たしていることが決定している
-	std::map<std::string, std::vector<std::string>> AcceptLanguageMap;
+	std::map<std::string, std::vector<std::string> > AcceptLanguageMap;
 	std::string::iterator itr = src.begin();
 	std::string::iterator last = src.end();
 	while (itr != last)
@@ -705,10 +705,10 @@ bool Response::isMatchAcceptCharsetFromat(std::string src)
 	return (true);
 }
 
-std::map<std::string, std::vector<std::string>> Response::parseAcceptCharset(std::string src)
+std::map<std::string, std::vector<std::string> > Response::parseAcceptCharset(std::string src)
 {
 	//ここに入ってくる時点で、Accept-Languageの形式は満たしていることが決定している
-	std::map<std::string, std::vector<std::string>> AcceptLanguageMap;
+	std::map<std::string, std::vector<std::string> > AcceptLanguageMap;
 	std::string::iterator itr = src.begin();
 	std::string::iterator last = src.end();
 	while (itr != last)
@@ -806,7 +806,7 @@ bool Response::isRedirection()
 	while (start != last)
 	{
 		std::string key = start->first;
-		int place = AbsolutePath.find(key);
+		size_t place = AbsolutePath.find(key);
 		if (place != std::string::npos)
 		{
 			return true;
@@ -1267,8 +1267,8 @@ std::string Response::GetSerachAbsolutePath()
 
 int Response::isLanguageFileExist(std::string SerachFileAbsolutePath)
 {
-	std::map<std::string, std::vector<std::string>>::reverse_iterator first = AcceptLanguageMap.rbegin();
-	std::map<std::string, std::vector<std::string>>::reverse_iterator last = AcceptLanguageMap.rend();
+	std::map<std::string, std::vector<std::string> >::reverse_iterator first = AcceptLanguageMap.rbegin();
+	std::map<std::string, std::vector<std::string> >::reverse_iterator last = AcceptLanguageMap.rend();
 	while (first != last)
 	{
 		int statusNo;
@@ -1309,8 +1309,8 @@ void Response::setContentLanguage()
 
 int Response::isCharsetFileExist(std::string SerachFileAbsolutePath)
 {
-	std::map<std::string, std::vector<std::string>>::reverse_iterator first = AcceptCharsetMap.rbegin();
-	std::map<std::string, std::vector<std::string>>::reverse_iterator last = AcceptCharsetMap.rend();
+	std::map<std::string, std::vector<std::string> >::reverse_iterator first = AcceptCharsetMap.rbegin();
+	std::map<std::string, std::vector<std::string> >::reverse_iterator last = AcceptCharsetMap.rend();
 	while (first != last)
 	{
 		int statusNo;
@@ -1343,8 +1343,8 @@ int Response::isCharsetFileExist(std::string SerachFileAbsolutePath)
 
 int Response::isCharsetAndLanguageFileExist(std::string SerachFileAbsolutePath)
 {
-	std::map<std::string, std::vector<std::string>>::reverse_iterator Cfirst = AcceptCharsetMap.rbegin();
-	std::map<std::string, std::vector<std::string>>::reverse_iterator Clast = AcceptCharsetMap.rend();
+	std::map<std::string, std::vector<std::string> >::reverse_iterator Cfirst = AcceptCharsetMap.rbegin();
+	std::map<std::string, std::vector<std::string> >::reverse_iterator Clast = AcceptCharsetMap.rend();
 
 	while (Cfirst != Clast)
 	{
@@ -1356,8 +1356,8 @@ int Response::isCharsetAndLanguageFileExist(std::string SerachFileAbsolutePath)
 		{
 			targetFileWithCharset = SerachFileAbsolutePath + "." + Charset[i]; // Charsetつけた
 
-			std::map<std::string, std::vector<std::string>>::reverse_iterator Lfirst = AcceptLanguageMap.rbegin();
-			std::map<std::string, std::vector<std::string>>::reverse_iterator Llast = AcceptLanguageMap.rend();
+			std::map<std::string, std::vector<std::string> >::reverse_iterator Lfirst = AcceptLanguageMap.rbegin();
+			std::map<std::string, std::vector<std::string> >::reverse_iterator Llast = AcceptLanguageMap.rend();
 			while (Lfirst != Llast)
 			{
 				std::vector<std::string> Languages = Lfirst->second;
@@ -1673,8 +1673,8 @@ bool Response::isLanguageFile(std::string FilePath, std::string fileExtention)
 		return (false);
 
 	//LanguageFile だったら、AcceptLanguageMapに該当する奴がいないかチェックしていって、あったらtrueを返す
-	std::map<std::string, std::vector<std::string>>::reverse_iterator first = AcceptLanguageMap.rbegin();
-	std::map<std::string, std::vector<std::string>>::reverse_iterator last = AcceptLanguageMap.rend();
+	std::map<std::string, std::vector<std::string> >::reverse_iterator first = AcceptLanguageMap.rbegin();
+	std::map<std::string, std::vector<std::string> >::reverse_iterator last = AcceptLanguageMap.rend();
 	while (first != last)
 	{
 		std::vector<std::string> values = first->second;
@@ -1697,8 +1697,8 @@ bool Response::isCharsetFile(std::string FilePath, std::string fileExtention)
 		return (false);
 
 	//LanguageFile だったら、AcceptLanguageMapに該当する奴がいないかチェックしていって、あったらtrueを返す
-	std::map<std::string, std::vector<std::string>>::reverse_iterator first = AcceptCharsetMap.rbegin();
-	std::map<std::string, std::vector<std::string>>::reverse_iterator last = AcceptCharsetMap.rend();
+	std::map<std::string, std::vector<std::string> >::reverse_iterator first = AcceptCharsetMap.rbegin();
+	std::map<std::string, std::vector<std::string> >::reverse_iterator last = AcceptCharsetMap.rend();
 	while (first != last)
 	{
 		std::vector<std::string> values = first->second;
